+++ pcap/src/lib.rs
@@ -1,8 +1,7 @@
-#[cfg(feature = "libpcap")]
-#[path = "./libpcap.rs"]
-mod backend;
+pub use libpcap::{NetworkInterface, PcapError, open};
 
-pub use backend::PcapError;
+#[cfg(feature = "libpcap")]
+mod libpcap;
 
 pub trait DataLinkSender {
     fn send_bytes(&mut self, buf: &[u8]) -> Result<(), PcapError>;
@@ -19,40 +18,4 @@ pub trait DataLinkReceiver {
 
 pub enum Channel {
     Ethernet(Box<dyn DataLinkSender>, Box<dyn DataLinkReceiver>),
-}
-
-pub fn open(ni: &NetworkInterface, promisc: bool) -> Result<Channel, PcapError> {
-    backend::open(ni, promisc)
-}
-
-#[derive(Debug, Clone)]
-pub struct NetworkInterface {
-    name: String,
-    description: String,
-    index: u32,
-}
-impl NetworkInterface {
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    pub fn description(&self) -> &str {
-        &self.description
-    }
-
-    pub fn index(&self) -> u32 {
-        self.index
-    }
-
-    pub fn list() -> Result<Vec<Self>, PcapError> {
-        backend::NetworkInterfaceInner::list()
-    }
-
-    pub fn find_by_name<S: AsRef<str>>(name: S) -> Option<Self> {
-        backend::NetworkInterfaceInner::find_by_name(name)
-    }
-
-    pub fn open(&self, promisc: bool) -> Result<Channel, PcapError> {
-        backend::open(self, promisc)
-    }
 }
